Fix-Missing dtype in array needed for py3.10

+++ Leaf_Segmentation_Functions_py3.py
@@ -1101,7 +1101,7 @@ def Load_Resize_and_Save_Stack(filepath, stack_name, rescale_factor, threshold_r
             print("***RESIZING***")
             #Iterating over each slice is faster than doing it in one call with transform.resize
             resized_shape = np.array(stack.shape)/np.array([1, rescale_factor, rescale_factor])
-            stack_rs = np.empty(shape = resized_shape.astype(np.int64))
+            stack_rs = np.empty(shape = resized_shape.astype(np.int64), dtype=np.int64)
             for idx in np.arange(stack_rs.shape[0]):
                 stack_rs[idx] = transform.resize(
                     stack[idx], [stack.shape[1]/rescale_factor, stack.shape[2]/rescale_factor],
